MicroFactor.Data

Description

Common data structures

Documentation

data TaggedValue r

The values that can be used in microfactor computations

Constructors

Boolean Bool

Integer Word

PortAddress Word

Instruction (MicroFactorInstruction r)

data MicroFactorInstruction r

A data type for instructions to be interpreted. It is generic over the type of values to be called

Constructors

Comment String
code annotation, does nothing

LiteralString String
debug output

Literal (TaggedValue r)
put value on stack

Call r
call other things

Operator MicroFactorOperator
operate on the stack(s)

data MicroFactorOperator

The basic opcodes of the language

Constructors

Execute

Debugger

ControlIte

ControlIf

ControlDo

ControlWhen

ControlUnless

ControlForever

ControlLoop

ControlWhile

ControlUntil

ControlDip

ControlKeep

LogicNot

LogicNor

LogicLt

LogicGt

LogicXor

LogicNand

LogicAnd

LogicXnor

LogicLte

LogicGte

LogicOr

StackNip

StackDrop

StackPick

StackDuplicate

StackOver

StackTuck

StackRoll

StackSwap

StackRotate

ArithAdd

ArithSub

ArithMul

ArithDiv

ArithMod

ArithDivmod

ArithAbs

ArithMax

ArithMin

Send

class InstructionRef a where

A typeclass for references to instructions so that the interpreter might be generic

Minimal complete definition

makeRef, resolveRef

Methods

makeRef :: [MicroFactorInstruction a] -> a

create a reference for anonymous blocks

resolveRef :: a -> [MicroFactorInstruction a]

get the referenced list of instruction

refName :: a -> String

refer to the instructions by name, instead of treating them as anonymous

data Nested

the most simple implementation of InstructionRef while avoiding infinite types