MicroFactor.Interpreter

Description

Runtime and interpreter for MicroFactor bytecode

Documentation

interpret :: InstructionRef a => [MicroFactorInstruction a] -> Thread a -> InterpreterResult a ()

Interpret a sequence of instructions on a thread. Creates output and affects the thread, and can fail with an error.

data InterpreterError

Things that can go wrong during interpretion

Constructors

StackUnderflow
an attempt was made to pop from the empty (data) stack

InvalidExecutionToken
an operation required an Instruction but was given some other value

TypeError String
some other operation did get the wrong operands

data ThreadState

Possible state for cooperative threads (not yet implemented)

Constructors

Running

Delayed

Yielded

Waiting

data InterpreterResult r a

The result of executing instructions

Constructors

InterpreterResult

Fields

interpreterOutput :: [String]

debug and Send output from the execution

interpreterThread :: Thread r

the new state of the thread

interpreterValue :: Either InterpreterError a

the result value of the current step

data Thread r

A structure representing a cooperative thread (that might be suspended)

Constructors

Thread

Fields

state :: ThreadState

dataStack :: [TaggedValue r]

returnStack :: [r]

newThread :: Thread r

A fresh new thread, ready to run instructions

showValue :: TaggedValue a -> String

shortened representation of TaggedValues