MicroFactor.Parser

Description

Parse string to bytecode and resolve identifiers to other functions

Documentation

data ParsedRef

The references to callables that the parser emits

Constructors

Anonymous [MicroFactorInstruction ParsedRef]
a parenthised code block

Named SourcePos String
an identifier

data ResolvedRef

An optionally named reference to more instructions

Constructors

ResolvedRef String [MicroFactorInstruction ResolvedRef]

expressionParser :: Parser [MicroFactorInstruction ParsedRef]

A parser for an expression, i.e. an instruction sequence with at least one element

data Command

The various commands available on the REPL

Constructors

Quit

List

Define String [MicroFactorInstruction ParsedRef]

Evaluate [MicroFactorInstruction ParsedRef]

ShowDef String

Load (Maybe String)

Save (Maybe String)

builtinSymbols :: InstructionRef r => Map String (MicroFactorInstruction r)

A lookup map of all the available Operators from their names, including the two boolean literals

commandParser :: Parser [Command]

A parser for a sequence of commands

dictionaryParser :: Parser [(String, [MicroFactorInstruction ParsedRef])]

type MicroFactorScope = Map String [MicroFactorInstruction ResolvedRef]

A Map of available functions

formatScope :: MicroFactorScope -> String

show the definitions that created the scope, line by line

resolve :: MicroFactorScope -> [MicroFactorInstruction ParsedRef] -> Either ParseError [MicroFactorInstruction ResolvedRef]

Try to resolve the Named ParsedRefs in a MicroFactorInstruction sequence. Uses the builtinSymbols or a lookup in the supplied scope, emitting an unknown identifier message if it fails for any

define

Arguments

:: String
Name of the new function

-> [MicroFactorInstruction ParsedRef]
Body of the new function

-> MicroFactorScope

-> Either ParseError MicroFactorScope

Add a function definition to the scope. Uses resolve, but also supports recursive definitions to the new function itself. TODO: Does not yet update references when redefining an existing name

formatErrorMessages :: ParseError -> String

show the errorMessages of a ParseError only (not its errorPos). Otherwise very similar to the Show instance. Notice the leading linebreak!